+++ tests/urls.py
@@ -12,7 +12,12 @@ from django_altcha import AltchaChallengeView
 urlpatterns = [
     path(
         "altcha/challenge/",
-        AltchaChallengeView.as_view(max_number=100),
+        AltchaChallengeView.as_view(cost=100),
         name="altcha_challenge",
+    ),
+    path(
+        "altcha/challenge/defaults/",
+        AltchaChallengeView.as_view(),
+        name="altcha_challenge_defaults",
     ),
 ]